svg/struct/scripted/svg-getIntersectionList-002.svg

<svg xmlns="http://www.w3.org/2000/svg" xmlns:h="http://www.w3.org/1999/xhtml" id="root"><h:script src="/resources/testharness.js"/><h:script src="/resources/testharnessreport.js"/><script>const test = async_test(&apos;getIntersectionList() on clipped element&apos;); window.onload = test.step_func_done(() =&gt; { const testSVG = document.getElementById(&apos;root&apos;); const r = testSVG.createSVGRect(); r.x = 0; r.y = 0; r.width = 100; r.height = 99.9; assert_equals(testSVG.getIntersectionList(r, testSVG).length, 1); });</script><rect width="200" height="200" x="0" y="0"/><clipPath id="MyClip"><rect width="200" height="100" y="100"/></clipPath><rect width="200" height="200" x="0" y="0" fill="green" clip-path="url(#MyClip)"/></svg>
